Building A Simple AI Chatbot With Web Speech API And Node.js
To build the web app, we’re going to take three major steps:

Use the Web Speech API’s SpeechRecognition interface to listen to the user’s voice.
Send the user’s message to a commercial natural-language-processing API as a text string.
Once API.AI returns the response text back, use the SpeechSynthesis interface to give it a synthetic voice.